use 64 values

+++ Backend/src/PeoConServer/user/contact.go
@@ -3,16 +3,16 @@ package user
 const NAME_SIZE uint32 = 20
 
 type ContactInfo struct {
-	userID uint32
-	flag   uint32
+	userID uint64
+	flag   uint64
 	name   [NAME_SIZE]byte
 }
 
 type ContactCache struct {
-	user1 uint32
-	user2 uint32
-	flag1 uint32
-	flag2 uint32
+	user1 uint64
+	user2 uint64
+	flag1 uint64
+	flag2 uint64
 }
 
 // ContactsRing size 32 -> 64 -> 128 -> ...
